requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==4.2.20
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
redis==5.2.1
routing==0.2.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
tzdata==2025.2
zope.interface==7.2
